Add a clause for closing delimiters that do not end their line
When we have something like:
```
x = [3,
     4,
     5
    ]
```
where the opening delimiter does not terminate the line, we end up with content
aligned inside the opening delimiter.  When this happens, we want the closing
delimiter to match the indentation of the opening delimiter, not the indentation
of the line where the opening delimiter occurred.

This commit implements this behavior by checking if a delimiter is followed by
something and if so setting the indent to the column in which the opening
delimiter appeared.