Add a test of a jsx element inside parens
This goes with issue #1.  It looks like there's two problems.  The first is that
the closing paren isn't indented correctly (which this checks).  The second is
that the jsx isn't recognized correctly for indentation until the closing paren.
(But after the closing paren is in place it does indent correctly, so this test
doesn't show an error in the jsx part).